+++ android/src/main/kotlin/tech/blockmanic/flutter_temi/FlutterTemiPlugin.kt
@@ -1,39 +1,69 @@
 package tech.blockmanic.flutter_temi
 
+//import io.flutter.plugin.common.PluginRegistry.Registrar
+import android.app.Activity
+import android.content.Context
+import android.content.pm.PackageManager
+import com.robotemi.sdk.Robot
+import com.robotemi.sdk.TtsRequest
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry.Registrar
-import com.robotemi.sdk.*
-import io.flutter.plugin.common.EventChannel
-import com.robotemi.sdk.TtsRequest
+import io.flutter.plugin.common.PluginRegistry
+
+//Method Call Handler can do V1 V2,
+//FlutterPlugin Give me most of the things except Activity
+//ActivityAware give me Activity
+
+class FlutterTemiPlugin : MethodCallHandler, FlutterPlugin, ActivityAware {
 
 
-class FlutterTemiPlugin : MethodCallHandler  {
+    public lateinit var plugin: FlutterTemiPlugin
+
+    //    private lateinit var channel: MethodChannel
+    public lateinit var application_context: Context
+    public lateinit var activity: Activity
 
     private val robot: Robot = Robot.getInstance()
-    private val goToLocationStatusChangedImpl : GoToLocationStatusChangedImpl = GoToLocationStatusChangedImpl()
-    private val onBeWithMeStatusChangedImpl : OnBeWithMeStatusChangedImpl = OnBeWithMeStatusChangedImpl()
-    private val onLocationsUpdatedImpl : OnLocationsUpdatedImpl = OnLocationsUpdatedImpl()
-    private val nlpImpl : NlpImpl = NlpImpl()
-    private val onUserInteractionChangedImpl : OnUserInteractionChangedImpl = OnUserInteractionChangedImpl()
-    private val ttsListenerImpl : TtsListenerImpl = TtsListenerImpl()
-    private val wakeupWordListenerImpl : WakeupWordListenerImpl = WakeupWordListenerImpl()
-    private val onConstraintBeWithStatusListenerImpl : OnConstraintBeWithStatusListenerImpl = OnConstraintBeWithStatusListenerImpl()
+    private val goToLocationStatusChangedImpl: GoToLocationStatusChangedImpl = GoToLocationStatusChangedImpl()
+    private val onBeWithMeStatusChangedImpl: OnBeWithMeStatusChangedImpl = OnBeWithMeStatusChangedImpl()
+    private val onLocationsUpdatedImpl: OnLocationsUpdatedImpl = OnLocationsUpdatedImpl()
+    private val nlpImpl: NlpImpl = NlpImpl()
+    private val onUserInteractionChangedImpl: OnUserInteractionChangedImpl = OnUserInteractionChangedImpl()
+    private val ttsListenerImpl: TtsListenerImpl = TtsListenerImpl()
+    private val asrListenerImpl: ASRListenerImpl = ASRListenerImpl()
+
+    private val wakeupWordListenerImpl: WakeupWordListenerImpl = WakeupWordListenerImpl()
+    private val onConstraintBeWithStatusListenerImpl: OnConstraintBeWithStatusListenerImpl = OnConstraintBeWithStatusListenerImpl()
+
     //private val onTelepresenceStatusChangedListenerImpl : OnTelepresenceStatusChangedListenerImpl = OnTelepresenceStatusChangedListenerImpl()
     //private val onUsersUpdatedListenerImpl : OnUsersUpdatedListenerImpl = OnUsersUpdatedListenerImpl()
-    private val onPrivacyModeChangedListenerImpl : OnPrivacyModeChangedListenerImpl = OnPrivacyModeChangedListenerImpl()
-    private val onBatteryStatusChangedListenerImpl : OnBatteryStatusChangedListenerImpl = OnBatteryStatusChangedListenerImpl()
-    private val onDetectionStateChangedListenerImpl : OnDetectionStateChangedListenerImpl = OnDetectionStateChangedListenerImpl()
-    private val onRobotReadyListenerImpl : OnRobotReadyListenerImpl = OnRobotReadyListenerImpl()
+    private val onPrivacyModeChangedListenerImpl: OnPrivacyModeChangedListenerImpl = OnPrivacyModeChangedListenerImpl()
+    private val onBatteryStatusChangedListenerImpl: OnBatteryStatusChangedListenerImpl = OnBatteryStatusChangedListenerImpl()
+    private val onDetectionStateChangedListenerImpl: OnDetectionStateChangedListenerImpl = OnDetectionStateChangedListenerImpl()
+    private val onRobotReadyListenerImpl: OnRobotReadyListenerImpl = OnRobotReadyListenerImpl()
+
 
     companion object {
+
+        private lateinit var channel: MethodChannel
+
+        /// 保留旧版本的兼容
         @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val channel = MethodChannel(registrar.messenger(), "flutter_temi")
+        fun registerWith(registrar: PluginRegistry.Registrar) {
+            channel = MethodChannel(registrar.messenger(), "flutter_temi")
             val plugin = FlutterTemiPlugin()
             channel.setMethodCallHandler(plugin)
+
+            //added
+            plugin.application_context = registrar.activity().getApplication()
+            plugin.activity = registrar.activity()
+
 
             val onBeWithMeEventChannel = EventChannel(registrar.messenger(), OnBeWithMeStatusChangedImpl.STREAM_CHANNEL_NAME)
             onBeWithMeEventChannel.setStreamHandler(plugin.onBeWithMeStatusChangedImpl)
@@ -49,6 +79,9 @@ class FlutterTemiPlugin : MethodCallHandler  {
 
             val ttsListenerEventChannel = EventChannel(registrar.messenger(), TtsListenerImpl.STREAM_CHANNEL_NAME)
             ttsListenerEventChannel.setStreamHandler(plugin.ttsListenerImpl)
+
+            val asrListenerEventChannel = EventChannel(registrar.messenger(), ASRListenerImpl.STREAM_CHANNEL_NAME)
+            asrListenerEventChannel.setStreamHandler(plugin.asrListenerImpl)
 
             val wakeupWordListenerEventChannel = EventChannel(registrar.messenger(), WakeupWordListenerImpl.STREAM_CHANNEL_NAME)
             wakeupWordListenerEventChannel.setStreamHandler(plugin.wakeupWordListenerImpl)
@@ -83,6 +116,7 @@ class FlutterTemiPlugin : MethodCallHandler  {
         robot.addNlpListener(this.nlpImpl)
         robot.addOnUserInteractionChangedListener(this.onUserInteractionChangedImpl)
         robot.addTtsListener(this.ttsListenerImpl)
+        robot.addAsrListener(this.asrListenerImpl) //asr
         robot.addWakeupWordListener(this.wakeupWordListenerImpl)
         robot.addOnConstraintBeWithStatusChangedListener(this.onConstraintBeWithStatusListenerImpl)
         //robot.addOnTelepresenceStatusChangedListener(this.onTelepresenceStatusChangedListenerImpl)
@@ -91,6 +125,7 @@ class FlutterTemiPlugin : MethodCallHandler  {
         robot.addOnBatteryStatusChangedListener(this.onBatteryStatusChangedListenerImpl)
         robot.addOnDetectionStateChangedListener(this.onDetectionStateChangedListenerImpl)
         robot.addOnRobotReadyListener(this.onRobotReadyListenerImpl)
+
     }
 
 
@@ -122,6 +157,16 @@ class FlutterTemiPlugin : MethodCallHandler  {
                 val speech = call.arguments<String>()
                 val request = TtsRequest.create(speech, true)
                 robot.speak(request)
+                result.success(true)
+            }
+            "temi_speak_force" -> {
+                val speech = call.arguments<String>()
+                val request = TtsRequest.create(speech, false)
+                robot.speak(request)
+                result.success(true)
+            }
+            "temi_finishe_conversation" -> {
+                robot.finishConversation()
                 result.success(true)
             }
             "temi_goto" -> {
@@ -186,22 +231,30 @@ class FlutterTemiPlugin : MethodCallHandler  {
             }
             "temi_get_contacts" -> {
                 val contacts = robot.allContact
-                val maps = contacts.map { contact -> OnUsersUpdatedListenerImpl.contactToMap(contact)}
+                val maps = contacts.map { contact -> OnUsersUpdatedListenerImpl.contactToMap(contact) }
                 result.success(maps)
             }
             "temi_get_recent_calls" -> {
                 val recentCalls = robot.recentCalls
-                val recentCallMaps = recentCalls.map {
-                    call ->
-                        val callMap = HashMap<String, Any?>(4)
-                        callMap["callType"] = call.callType
-                        callMap["sessionId"] = call.sessionId
-                        callMap["timestamp"] = call.timestamp
-                        callMap["userId"] = call.userId
-                        callMap
+                val recentCallMaps = recentCalls.map { call ->
+                    val callMap = HashMap<String, Any?>(4)
+                    callMap["callType"] = call.callType
+                    callMap["sessionId"] = call.sessionId
+                    callMap["timestamp"] = call.timestamp
+                    callMap["userId"] = call.userId
+                    callMap
                 }
                 result.success(recentCallMaps)
             }
+            "temi_wakeup" -> {
+                robot.wakeup()
+                result.success(true)
+            }
+            "temi_showAppList" -> {
+                robot.showAppList()
+                result.success(true)
+            }
+
             "temi_toggle_wakeup" -> {
                 val disable = call.arguments<Boolean>()
                 robot.toggleWakeup(disable)
@@ -212,9 +265,91 @@ class FlutterTemiPlugin : MethodCallHandler  {
                 robot.toggleNavigationBillboard(hide)
                 result.success(true)
             }
+            "temi_turnKoiskMode" -> {
+                val activityInfo = application_context.getPackageManager()
+                        .getActivityInfo(activity.getComponentName(), PackageManager.GET_META_DATA)
+                robot.onStart(activityInfo)
+                result.success(true)
+            }
+            "temi_repose" -> {
+                robot.repose()
+                result.success(true)
+            }
+
             else -> {
                 result.notImplemented()
             }
         }
+    }
+
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        val channel = MethodChannel(binding.binaryMessenger, "flutter_temi")
+        plugin = FlutterTemiPlugin()
+        application_context = binding.applicationContext
+        channel.setMethodCallHandler(this)
+
+
+        val onBeWithMeEventChannel = EventChannel(binding.binaryMessenger, OnBeWithMeStatusChangedImpl.STREAM_CHANNEL_NAME)
+        onBeWithMeEventChannel.setStreamHandler(plugin.onBeWithMeStatusChangedImpl)
+
+        val onLocationStatusChangeEventChannel = EventChannel(binding.binaryMessenger, GoToLocationStatusChangedImpl.STREAM_CHANNEL_NAME)
+        onLocationStatusChangeEventChannel.setStreamHandler(plugin.goToLocationStatusChangedImpl)
+
+        val onLocationsUpdatedEventChannel = EventChannel(binding.binaryMessenger, OnLocationsUpdatedImpl.STREAM_CHANNEL_NAME)
+        onLocationsUpdatedEventChannel.setStreamHandler(plugin.onLocationsUpdatedImpl)
+
+        val onNlpEventChannel = EventChannel(binding.binaryMessenger, NlpImpl.STREAM_CHANNEL_NAME)
+        onNlpEventChannel.setStreamHandler(plugin.nlpImpl)
+
+        val ttsListenerEventChannel = EventChannel(binding.binaryMessenger, TtsListenerImpl.STREAM_CHANNEL_NAME)
+        ttsListenerEventChannel.setStreamHandler(plugin.ttsListenerImpl)
+
+        val asrListenerEventChannel = EventChannel(binding.binaryMessenger, ASRListenerImpl.STREAM_CHANNEL_NAME)
+        asrListenerEventChannel.setStreamHandler(plugin.asrListenerImpl)
+
+        val wakeupWordListenerEventChannel = EventChannel(binding.binaryMessenger, WakeupWordListenerImpl.STREAM_CHANNEL_NAME)
+        wakeupWordListenerEventChannel.setStreamHandler(plugin.wakeupWordListenerImpl)
+
+        val onConstraintBeWithStatusListenerEventChannel = EventChannel(binding.binaryMessenger, OnConstraintBeWithStatusListenerImpl.STREAM_CHANNEL_NAME)
+        onConstraintBeWithStatusListenerEventChannel.setStreamHandler(plugin.onConstraintBeWithStatusListenerImpl)
+
+        //val onTelepresenceStatusChangedListenerEventChannel = EventChannel(registrar.messenger(), OnTelepresenceStatusChangedListenerImpl.STREAM_CHANNEL_NAME)
+        //onTelepresenceStatusChangedListenerEventChannel.setStreamHandler(plugin.onTelepresenceStatusChangedListenerImpl)
+
+        //val onUsersUpdatedListenerEventChannel = EventChannel(registrar.messenger(), OnUsersUpdatedListenerImpl.STREAM_CHANNEL_NAME)
+        //onUsersUpdatedListenerEventChannel.setStreamHandler(plugin.onUsersUpdatedListenerImpl)
+
+        val onPrivacyModeChangedListenerEventChannel = EventChannel(binding.binaryMessenger, OnPrivacyModeChangedListenerImpl.STREAM_CHANNEL_NAME)
+        onPrivacyModeChangedListenerEventChannel.setStreamHandler(plugin.onPrivacyModeChangedListenerImpl)
+
+        val onBatteryStatusChangedListenerEventChannel = EventChannel(binding.binaryMessenger, OnBatteryStatusChangedListenerImpl.STREAM_CHANNEL_NAME)
+        onBatteryStatusChangedListenerEventChannel.setStreamHandler(plugin.onBatteryStatusChangedListenerImpl)
+
+        val onDetectionStateChangedEventChannel = EventChannel(binding.binaryMessenger, OnDetectionStateChangedListenerImpl.STREAM_CHANNEL_NAME)
+        onDetectionStateChangedEventChannel.setStreamHandler(plugin.onDetectionStateChangedListenerImpl)
+
+        val onRobotReadyEventChannel = EventChannel(binding.binaryMessenger, OnRobotReadyListenerImpl.STREAM_CHANNEL_NAME)
+        onRobotReadyEventChannel.setStreamHandler(plugin.onRobotReadyListenerImpl)
+    }
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        activity = binding.activity
+
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        activity = binding.activity
+    }
+
+    override fun onDetachedFromActivity() {
+
     }
 }