+++ CHANGELOG.md
@@ -1,5 +1,23 @@
-## 0.0.1
+## 0.0.13
+-- Make V2 Binding
 
+## 0.0.12
+-- Upgraded with the Temi SDK
+-- Add support of the repose
+
+
+
+## 0.0.9
+-- Fix the location Stream
+
+
+## 0.0.8
+August 16, 2020 4:22 PM
+Upgraded Android X, Mix of V1 and V2 Embedding
+
+Interesting!!
+
+## 0.0.1
 10 November 2019 6:00 PM 
 - Most basic Features works, merge with